services/plesk-mail-api/requirements.txt
requests>=2.32,<3
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-dotenv>=1.0,<2
